+++ app/tabs/maps.tsx
@@ -395,7 +395,7 @@ async function getMapsTabContents(tab: TabManagerTab): Promise<JSX.Element> {
             if (!tablesContainerRef.current) return;
             if (reloadData) {
                 mode = config.views.maps.mode;
-                console.log(query);
+                console.debug(query);
                 tablesContents = await Promise.all(
                     ConfigConstants.views.Maps.mapsTabModeToSectionIDs[mode].map(
                         async (sectionID: (typeof ConfigConstants.views.Maps.mapsTabModeToSectionIDs)[typeof mode][number]): Promise<JSX.Element[]> =>
@@ -516,7 +516,7 @@ async function getMapsTabContents(tab: TabManagerTab): Promise<JSX.Element> {
                                     queryData[key].push(condition.value);
                                 }
                                 const textQueryData = searchString.getTextSegments();
-                                console.log(searchString, queryData, textQueryData);
+                                console.debug("Search query:", searchString, queryData, textQueryData);
                                 if (lastHideErrorPopupFunction) lastHideErrorPopupFunction();
                                 function showError(options: { message: string }): void {
                                     if (!searchRefs.searchTextBox.current || !searchRefs.searchTextBoxErrorPopup.current) return;