+++ app/tabs/none.tsx
@@ -4,7 +4,7 @@ import _React from "preact/compat";
 export default function NoneTab(): JSX.SpecificElement<"h1"> {
     return (
         <center>
-            <h2>Select a tab from the left sidebar to get started.</h2>
+            <h2 class="nsel">Select a tab from the left sidebar to get started.</h2>
         </center>
     );
 }