+++ app/tabs/structures.tsx
@@ -309,7 +309,7 @@ async function getStructuresTabContents(tab: TabManagerTab): Promise<JSX.Element
             if (!tablesContainerRef.current) return;
             if (reloadData) {
                 mode = config.views.structures.mode;
-                console.log(query);
+                console.debug(query);
                 tablesContents = await Promise.all(
                     ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode].map(
                         (sectionID: (typeof ConfigConstants.views.Structures.structuresTabModeToSectionIDs)[typeof mode][number]): Promise<JSX.Element[]> =>
@@ -803,7 +803,7 @@ async function getStructuresTabContents(tab: TabManagerTab): Promise<JSX.Element
                                     queryData[key].push(condition.value);
                                 }
                                 const textQueryData = searchString.getTextSegments();
-                                console.log(searchString, queryData, textQueryData);
+                                console.debug("Search query:", searchString, queryData, textQueryData);
                                 if (lastHideErrorPopupFunction) lastHideErrorPopupFunction();
                                 function showError(options: { message: string }): void {
                                     if (!searchRefs.searchTextBox.current || !searchRefs.searchTextBoxErrorPopup.current) return;