+++ app/tabs/tickingAreas.tsx
@@ -366,7 +366,7 @@ async function getTickingAreasTabContents(tab: TabManagerTab): Promise<JSX.Eleme
             if (!tablesContainerRef.current) return;
             if (reloadData) {
                 mode = config.views.tickingAreas.mode;
-                console.log(query);
+                console.debug(query);
                 tablesContents = await Promise.all(
                     ConfigConstants.views.TickingAreas.tickingAreasTabModeToSectionIDs[mode].map(
                         async (
@@ -489,7 +489,7 @@ async function getTickingAreasTabContents(tab: TabManagerTab): Promise<JSX.Eleme
                                     queryData[key].push(condition.value);
                                 }
                                 const textQueryData = searchString.getTextSegments();
-                                console.log(searchString, queryData, textQueryData);
+                                console.debug("Search query:", searchString, queryData, textQueryData);
                                 if (lastHideErrorPopupFunction) lastHideErrorPopupFunction();
                                 function showError(options: { message: string }): void {
                                     if (!searchRefs.searchTextBox.current || !searchRefs.searchTextBoxErrorPopup.current) return;